use shared Button for password toggle, move auth types to auth.type.t
