+++ apps/web/src/services/auth.service.ts
@@ -1,21 +1,8 @@
 import { endpoints } from "@/data/endpoints";
 import { ErrorApiResponse, Response } from "@/types/api.type";
+import { AdminLoginData, AdminLoginPayload } from "@/types/auth.type";
 import { axiosPost } from "@/utils/api";
-
-export interface AuthTokens {
-  access_token: string;
-  refresh_token: string;
-}
-
-export interface AdminLoginPayload {
-  email: string;
-  password: string;
-}
-
-export interface AdminLoginData {
-  token: AuthTokens;
-}
-
+ 
 export class AuthService {
   static async login(payload: unknown) {
     try {
@@ -23,27 +10,28 @@ export class AuthService {
         endpoints.auth.login,
         payload,
       );
-
+ 
       if (!("data" in response)) {
         return response as ErrorApiResponse;
       }
-
+ 
       return response;
     } catch (error) {
       throw error;
     }
   }
-
+ 
   static async adminLogin(payload: AdminLoginPayload) {
     const response = await axiosPost<Response<AdminLoginData>>(
       endpoints.auth.adminLogin,
       payload,
     );
-
+ 
     if (!("data" in response)) {
       return response as ErrorApiResponse;
     }
-
+ 
     return response;
   }
 }
+
