+++ apps/web/src/types/button.type.ts
@@ -1,7 +1,7 @@
 import { ButtonHTMLAttributes, ReactNode } from "react";
 
 export interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
-  text: string;
+  text?: string;
   icon?: ReactNode;
   loading?: boolean;
   iconPosition?: "left" | "right";
